// src/app/components/pages/signup/interfaces/signup.interfaces.ts
export interface SignupForm {
    firstName: string;
    lastName: string;
    phoneNumber: string;
    email: string;
    roleName: string;
    password: string;
  }
  
  export interface ProfileData {
    firstName: string;
    lastName: string;
    phoneNumber: string;
    bio: string;
    designation: string;
  }